+++ src/components/BaseInput.vue
@@ -24,7 +24,6 @@
 </template>
 
 <script setup>
-import { defineProps, defineEmits } from 'vue';
 
 const props = defineProps({
   id: {